Move certs.go from signedexchange to internal
