+++ go/signedexchange/signedexchange_test.go
@@ -62,12 +62,12 @@ func testForEachVersion(t *testing.T, testFunc func(ver version.Version, t *test
 }
 
 func TestSignedExchange(t *testing.T) {
-	certs, err := ParseCertificates([]byte(pemCerts))
+	certs, err := signingalgorithm.ParseCertificates([]byte(pemCerts))
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	privKey, err := ParsePrivateKey([]byte(pemPrivateKey))
+	privKey, err := signingalgorithm.ParsePrivateKey([]byte(pemPrivateKey))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -187,10 +187,10 @@ func TestSignedExchangeBannedCertUrlScheme(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		certs, _ := ParseCertificates([]byte(pemCerts))
+		certs, _ := signingalgorithm.ParseCertificates([]byte(pemCerts))
 		certUrl, _ := url.Parse("http://example.com/cert.msg")
 		validityUrl, _ := url.Parse("https://example.com/resource.validity")
-		privKey, _ := ParsePrivateKey([]byte(pemPrivateKey))
+		privKey, _ := signingalgorithm.ParsePrivateKey([]byte(pemPrivateKey))
 		s := &Signer{
 			Date:        signatureDate,
 			Expires:     signatureDate.Add(1 * time.Hour),
@@ -214,12 +214,12 @@ func createTestExchange(ver version.Version, t *testing.T) (e *Exchange, s *Sign
 		t.Fatal(err)
 	}
 
-	certs, err := ParseCertificates([]byte(pemCerts))
+	certs, err := signingalgorithm.ParseCertificates([]byte(pemCerts))
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	privKey, err := ParsePrivateKey([]byte(pemPrivateKey))
+	privKey, err := signingalgorithm.ParsePrivateKey([]byte(pemPrivateKey))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -358,7 +358,7 @@ func TestVerifyBadSignature(t *testing.T) {
 func TestVerifyNoContentType(t *testing.T) {
 	testForEachVersion(t, func(ver version.Version, t *testing.T) {
 		e, s, c := createTestExchange(ver, t)
-		e.ResponseHeaders.Del("Content-Type");
+		e.ResponseHeaders.Del("Content-Type")
 		if err := e.AddSignatureHeader(s); err != nil {
 			t.Fatal(err)
 		}
